refactor(Backend)!: Organize schemas to separate files
- Moved various schemas into dedicated files: test_item_schemas.py, test_instance_schemas.py, section_schemas.py, student_schemas.py, and student_answer_schemas.py.
- Updated __init__.py to import these new schema modules, streamlining the schema management process.
- This restructuring enhances code organization and readability.
