+++ app/backend/schemas/__init__.py
@@ -1,212 +1,5 @@
-from pydantic import BaseModel, Field
-from typing import List, Optional
-
-
-
-# ==============================
-#region Test Item Schemas
-class TestItemSummary(BaseModel):
-    """Response model for GET /items endpoint (minimal fields per contract)"""
-    item_id: int
-    label: str
-    question: str
-    is_problem_solving: bool
-    expected_answer_rubric_questions: str
-
-class TestItemsResponse(BaseModel):
-    test_id: str
-    items: List[TestItemSummary]
-
-class NewTestItemRequest(BaseModel):
-    """Request body for POST /items (item_id is generated server-side)"""
-    label: str = "Item"
-    question: str = "Untitled Question"
-    is_problem_solving: bool = False
-    expected_answer_rubric_questions: str = ""
-
-class NewTestItemResponse(BaseModel):
-    """Response for POST /items (returns full created items)"""
-    items: List["FullTestItemResponse"]
-
-class UpdateTestItemRequest(BaseModel):
-    """Request body for PATCH /items/{item_id} (partial updates)"""
-    label: Optional[str] = Field(None, min_length=1)
-    question: Optional[str] = Field(None, min_length=1)
-    is_problem_solving: Optional[bool] = None
-    expected_answer_rubric_questions: Optional[str] = Field(None, min_length=1)
-
-class FullTestItemResponse(BaseModel):
-    """Full item representation for PATCH response and POST response"""
-    item_id: int
-    label: str
-    question: str
-    is_problem_solving: bool
-    expected_answer_rubric_questions: str
-
-class TestPaperResponse(BaseModel):
-    paper_id: int
-    test_id: str
-    student_no: str
-    is_done_rendering: bool
-#endregion
-# ==============================
-
-
-
-# ==============================
-#region Test Instance Schemas
-class NewTestInstanceRequest(BaseModel):
-    """Request body for POST /test_instances"""
-    name: str
-    section_id: int
-    date: str
-
-class UpdateTestItemSchema(BaseModel):
-    """Schema for updating test items within a test instance"""
-    item_id: int
-    label: str = Field(..., min_length=1)
-    question: Optional[str] = Field(None, min_length=1)
-    is_problem_solving: Optional[bool] = None
-    expected_answer_rubric_questions: str = Field(..., min_length=1)
-
-class UpdateTestInstanceRequest(BaseModel):
-    """Request body for PATCH /test_instances/{test_id}"""
-    date: Optional[str] = Field(None, min_length=1)
-    items: Optional[List[UpdateTestItemSchema]] = None
-    
-class TestInstanceResponse(BaseModel):
-    """Full test instance response (used in PATCH and future GET with items)"""
-    name: str
-    section_id: int
-    date: str
-    test_id: str
-    is_done_rendering: bool
-    items: List[TestItemSummary]
-#endregion
-# ==============================
-
-
-
-# ==============================
-#region Section schemas
-class SectionCreate(BaseModel):
-    section: str
-
-class SectionResponse(BaseModel):
-    section_id: int
-    section_name: str
-#endregion
-# ==============================
-
-
-
-# ==============================
-#region Student Schemas
-class StudentResponse(BaseModel):
-    student_no: str
-    name: str
-    section_id: int
-
-class NewStudentRequest(BaseModel):
-    student_no: str
-    name: str
-    section_id: int
-
-class UpdateStudentRequest(BaseModel):
-    name: Optional[str] = None
-    section_id: Optional[int] = None
-#endregion
-# ==============================
-
-
-
-# ==============================
-#region Student answer schemas
-class StudentAnswerRequest(BaseModel):
-    paper_id: int
-    item_id: int
-    image_directory: str
-    ai_evaluation: Optional[str] = ""
-    is_done_rendering: bool = False
-
-class StudentAnswerResponse(BaseModel):
-    answer_id: int
-    paper_id: int
-    item_id: int
-    image_directory: str
-    ai_evaluation: str
-    is_done_rendering: bool
-    scores: str
-
-class StudentAnswerSummary(BaseModel):
-    student_no: str
-    name: str
-    item_id: int
-    label: str
-    image_directory: str
-    ai_evaluation: str
-    is_done_rendering: bool
-
-class BoxesItem(BaseModel):
-    index: int
-    image_directory: str
-    item_number: str
-
-class CommitBoxesRequest(BaseModel):
-    boxes: list[BoxesItem]
-
-class ImagePreprocessResponse(BaseModel):
-    num_boxes: int
-    boxes: list[BoxesItem]
-
-class LabelSaveBoxesResponse(BaseModel):
-    num_boxes: int
-    boxes: list[BoxesItem]
-
-class UpdateSegmentationResponse(BaseModel):
-    image_directory: str
-
-class StudentStatusItem(BaseModel):
-    student_no: str
-    name: str
-    is_done_rendering: bool
-    total_score: str
-
-class TestStatusesResponse(BaseModel):
-    test_id: str
-    statuses: list[StudentStatusItem]
-
-class AIEvaluationItem(BaseModel):
-    item_id: int
-    answer_id: int
-    label: str
-    question: str
-    expected_answer_rubric_questions: str
-    ai_evaluation: str
-
-class AIEvaluationItemWithScores(BaseModel):
-    item_id: int
-    answer_id: int
-    label: str
-    question: str
-    expected_answer_rubric_questions: str
-    ai_evaluation: str
-    scores: str
-
-class StudentEvaluationsStore(BaseModel):
-    student_no: str
-    name: str
-    evaluations: list[AIEvaluationItem]
-
-class TestResultsResponse(BaseModel):
-    test_id: str
-    students: list[StudentEvaluationsStore]
-
-class StudentResultsResponse(BaseModel):
-    test_id: str
-    student_no: str
-    name: str
-    evaluations: list[AIEvaluationItemWithScores]
-#endregion
-# ==============================
-
+from .test_item_schemas import *
+from .test_instance_schemas import *
+from .section_schemas import *
+from .student_schemas import *
+from .student_answer_schemas import *
