+++ app/backend/schemas/section_schemas.py
@@ -1,0 +1,10 @@
+from pydantic import BaseModel
+
+
+
+class SectionCreate(BaseModel):
+    section: str
+
+class SectionResponse(BaseModel):
+    section_id: int
+    section_name: str
